CMPE-242: Lab3: Initial program
This is the initial program for CMPE-242 lab3 (PID controller). It has
a few modules/functions as the professor requires. However, lots of parts
should be not on the right track and needs to be refined as his latest
handout on the Yahoo Group, including the existing TODO and FIXME in the
program.

Signed-off-by: Nicolin Chen <nicoleotsuka@gmail.com>